+++ source/compress_integer_prn_512_carryover.cpp
@@ -235,7 +235,7 @@ namespace JASS
 
 		while (1)
 			{
-			width = find_first_set_bit(selector);
+			width = (uint32_t)find_first_set_bit(selector);
 			mask = _mm256_loadu_si256((__m256i *)mask_set[width]);
 			_mm256_storeu_si256(into, _mm256_and_si256(payload1, mask));
 			_mm256_storeu_si256(into + 1, _mm256_and_si256(payload2, mask));
@@ -267,7 +267,7 @@ namespace JASS
 				/*
 					get the low bits and write to memory
 				*/
-				width = find_first_set_bit(selector);
+				width = (uint32_t)find_first_set_bit(selector);
 
 				high_bits1 = _mm256_slli_epi32(high_bits1, width);
 				high_bits2 = _mm256_slli_epi32(high_bits2, width);
