+++ source/compress_integer_simple_9.cpp
@@ -47,7 +47,7 @@ namespace JASS
 		--------------------------------------
 		The number of bits that simple-9 will be used to store an integer of the given the number of bits in length
 	*/
-	const size_t compress_integer_simple_9::bits_to_use[] =
+	const uint32_t compress_integer_simple_9::bits_to_use[] =
 		{
 		 1,  1,  2,  3,  4,  5,  7,  7, 
 		 9,  9, 14, 14, 14, 14, 14, 28, 
@@ -64,7 +64,7 @@ namespace JASS
 		------------------------------------
 		The row of the table to use given the number of integers we can pack into the word
 	*/
-	const size_t compress_integer_simple_9::table_row[] =
+	const uint32_t compress_integer_simple_9::table_row[] =
 		{
 		0, 1, 2, 3, 4, 4, 5, 5, 
 		6, 6, 6, 6, 6, 7, 7, 7, 
@@ -77,7 +77,7 @@ namespace JASS
 		----------------------------------------------
 		Number of bits to shift across when packing - is sum of prior packed ints
 	*/
-	const size_t compress_integer_simple_9::simple9_shift_table[] =
+	const uint32_t compress_integer_simple_9::simple9_shift_table[] =
 		{
 		0, 1, 2, 3, 4, 5, 6, 7, 8, 9, 10, 11, 12, 13, 14, 15, 16, 17, 18, 19, 20, 21, 22, 23, 24, 25, 26, 27,
 		0, 2, 4, 6, 8, 10, 12, 14, 16, 18, 20, 22, 24, 26, 28, 28, 28, 28, 28, 28, 28, 28, 28, 28, 28, 28, 28, 28,
@@ -95,7 +95,7 @@ namespace JASS
 		--------------------------------------------
 		Number of integers packed into a 32-bit word, given its mask type
 	*/
-	const size_t compress_integer_simple_9::ints_packed_table[] =
+	const uint32_t compress_integer_simple_9::ints_packed_table[] =
 		{
 		28, 14, 9, 7, 5, 4, 3, 2, 1
 		};
@@ -105,7 +105,7 @@ namespace JASS
 		-----------------------------------------
 		Bitmask map for valid masks at an offset (column) for some num_bits_needed (row).
 	*/
-	const size_t compress_integer_simple_9::can_pack_table[] =
+	const uint32_t compress_integer_simple_9::can_pack_table[] =
 		{
 		0x01ff, 0x00ff, 0x007f, 0x003f, 0x001f, 0x000f, 0x000f, 0x0007, 0x0007, 0x0003, 0x0003, 0x0003, 0x0003, 0x0003, 0x0001, 0x0001, 0x0001, 0x0001, 0x0001, 0x0001, 0x0001, 0x0001, 0x0001, 0x0001, 0x0001, 0x0001, 0x0001, 0x0001,
 		0x01fe, 0x00fe, 0x007e, 0x003e, 0x001e, 0x000e, 0x000e, 0x0006, 0x0006, 0x0002, 0x0002, 0x0002, 0x0002, 0x0002, 0x0000, 0x0000, 0x0000, 0x0000, 0x0000, 0x0000, 0x0000, 0x0000, 0x0000, 0x0000, 0x0000, 0x0000, 0x0000, 0x0000,
@@ -124,7 +124,7 @@ namespace JASS
 		---------------------------------------------------
 		We AND out masks for offsets where we don't know if we can fully pack for that offset
 	*/
-	const size_t compress_integer_simple_9::invalid_masks_for_offset[] =
+	const uint32_t compress_integer_simple_9::invalid_masks_for_offset[] =
 		{
 		0x0000, 0x0100, 0x0180, 0x01c0, 0x01e0, 0x01f0, 0x01f0, 0x01f8, 0x01f8, 0x01fc, 0x01fc, 0x01fc, 0x01fc, 0x01fc, 0x01fe, 0x01fe, 0x01fe, 0x01fe, 0x01fe, 0x01fe, 0x01fe, 0x01fe, 0x01fe, 0x01fe, 0x01fe, 0x01fe, 0x01fe, 0x01fe, 0x01ff
 		};
@@ -134,7 +134,7 @@ namespace JASS
 		----------------------------------------------
 		Translates the 'bits_needed' to the appropriate 'row' offset for use with can_pack table.
 	*/
-	const size_t compress_integer_simple_9::row_for_bits_needed[] =
+	const uint32_t compress_integer_simple_9::row_for_bits_needed[] =
 		{
 		0, 0, 28, 56, 84, 112, 140, 140, 168, 168, 196, 196, 196, 196, 196, 224, 224, 224, 224, 224, 224, 224, 224, 224, 224, 224, 224, 224, 224,			// Valid
 		252, 252, 252, 252, 252, 252, 252, 252, 252, 252, 252, 252, 252, 252, 252, 252, 252, 252, 252, 252, 252, 252, 252, 252, 252, 252, 252, 252, 252, 252, 252, 252, 252, 252, 252, 252		// Overflow
